added pie chart plot for annotation pretty names by type

+++ util/post_processing_build_methods.py
@@ -3,8 +3,9 @@
 from pathlib import Path
 
 import pandas as pd
-from post_processing import retrieve_pat_annots_mct_epr
 from tqdm import tqdm
+
+from pat2vec.util.post_processing import retrieve_pat_annots_mct_epr
 
 
 def filter_annot_dataframe(df, annot_filter_arguments):
@@ -60,25 +61,29 @@ def filter_annot_dataframe(df, annot_filter_arguments):
 # filtered_df = filter_annot_dataframe(df, annot_filter_arguments)
 
 
-def build_merged_epr_mct_annot_df(all_pat_list, config_obj):
+from pathlib import Path
+
+
+def build_merged_epr_mct_annot_df(all_pat_list, config_obj, overwrite=False):
     """
     Build a merged DataFrame containing annotations for multiple patients using MCT and EPR data.
 
     Parameters:
     - config_obj (ConfigObject): An object containing configuration settings, including project name,
                                 patient list, etc.
+    - overwrite (bool): If True, overwrite the existing output file. Default is False.
 
     Returns:
-    None
+    File path to output
 
     This function creates a directory for merged batches, retrieves annotations for each patient,
     and writes the merged annotations to a CSV file named 'annots_mct_epr.csv'.
-    If the output file already exists, subsequent batches are appended to it.
+    If the output file already exists and overwrite is False, subsequent batches are appended to it.
 
     Example usage:
     ```
     config = ConfigObject(...)  # Create or load your configuration object
-    build_merged_epr_mct_annot_df(config)
+    build_merged_epr_mct_annot_df(config, overwrite=True)
     ```
 
     """
@@ -87,17 +92,20 @@ def build_merged_epr_mct_annot_df(all_pat_list, config_obj):
 
     output_file_path = directory_path + 'annots_mct_epr.csv'
 
-    for i in tqdm(range(0, len(all_pat_list)), total=len(all_pat_list)):
-        current_pat_idcode = all_pat_list[i]
-        all_annots = retrieve_pat_annots_mct_epr(
-            current_pat_idcode, config_obj)
+    if not overwrite and Path(output_file_path).is_file():
+        print("Output file already exists. Appending to the existing file.")
+    else:
+        for i in tqdm(range(0, len(all_pat_list)), total=len(all_pat_list)):
+            current_pat_idcode = all_pat_list[i]
+            all_annots = retrieve_pat_annots_mct_epr(
+                current_pat_idcode, config_obj)
 
-        if i == 0:
-            # Create the output file and write the first batch directly
-            all_annots.to_csv(output_file_path, index=False)
-        else:
-            # Append each result to the output file
-            all_annots.to_csv(output_file_path, mode='a',
-                              header=False, index=False)
+            if i == 0:
+                # Create the output file and write the first batch directly
+                all_annots.to_csv(output_file_path, index=False)
+            else:
+                # Append each result to the output file
+                all_annots.to_csv(output_file_path, mode='a',
+                                  header=False, index=False)
 
     return output_file_path
